requirements.txt
dj-database-url==0.5.0
Django==2.1.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.5
whitenoise==4.1
psycopg2
social_django
